backend final

+++ information/models.py
@@ -143,7 +143,7 @@ class AB01(models.Model):
     date = models.CharField(max_length=8, blank=True, null=True, verbose_name='기준일자')
     rep_code = models.CharField(max_length=2, choices=BS_REPORT, verbose_name="보고서코드")
     item_code = models.CharField(max_length=4, verbose_name="항목코드")
-    price = models.IntegerField(blank=True, null=True, verbose_name="금액")
+    price = models.BigIntegerField(blank=True, null=True, verbose_name="금액")
     ratio = models.FloatField(blank=True, null=True, verbose_name='구성비')
     change_rate = models.FloatField(blank=True, null=True, verbose_name='증감율')
 
@@ -196,7 +196,7 @@ class AD01(models.Model):
     date = models.CharField(max_length=8, blank=True, null=True, verbose_name='기준일자')
     rep_code = models.CharField(max_length=2, choices=BS_REPORT, verbose_name="보고서코드")
     item_code = models.CharField(max_length=4, verbose_name="항목코드")
-    price = models.IntegerField(blank=True, null=True, verbose_name="금액")
+    price = models.BigIntegerField(blank=True, null=True, verbose_name="금액")
     ratio = models.FloatField(blank=True, null=True, verbose_name='구성비')
     change_rate = models.FloatField(blank=True, null=True, verbose_name='증감율')
